requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
Django==4.1.6
django-anymail==9.0
idna==3.4
pycparser==2.21
python-decouple==3.7
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
